learn(performance): exercise solution - code splitting'
