+++ src/app.exercise.js
@@ -1,16 +1,17 @@
 import * as React from 'react'
-import {useAuth} from './context/auth-context'
-// 🐨 you'll want to render the FullPageSpinner as the fallback
-// import {FullPageSpinner} from './components/lib'
+import { useAuth } from './context/auth-context'
+import { FullPageSpinner } from './components/lib'
 
-// 🐨 exchange these for React.lazy calls
-import {AuthenticatedApp} from './authenticated-app'
-import {UnauthenticatedApp} from './unauthenticated-app'
+const AuthenticatedApp = React.lazy(() => import('./authenticated-app'))
+const UnauthenticatedApp = React.lazy(() => import('./unauthenticated-app'))
 
 function App() {
-  const {user} = useAuth()
-  // 🐨 wrap this in a <React.Suspense /> component
-  return user ? <AuthenticatedApp /> : <UnauthenticatedApp />
+  const { user } = useAuth()
+  return (
+    <React.Suspense fallback={<FullPageSpinner />}>
+      {user ? <AuthenticatedApp /> : <UnauthenticatedApp />}
+    </React.Suspense>
+  )
 }
 
-export {App}
+export { App }
