+++ src/app.js
@@ -1,6 +1,6 @@
-export * from './app.final'
+// export * from './app.final'
 
-// export * from './app.exercise'
+export * from './app.exercise'
 
 // 💯 Prefetch the Authenticated App
 // export * from './app.extra-1'
