+++ src/authenticated-app.js
@@ -1,4 +1,6 @@
-export * from './authenticated-app.final'
-export {default} from './authenticated-app.final'
+// export * from './authenticated-app.final'
+// export {default} from './authenticated-app.final'
 
-// export * from './authenticated-app.exercise'
+export * from './authenticated-app.exercise'
+export {default} from './authenticated-app.exercise'
+
