+++ src/unauthenticated-app.exercise.js
@@ -108,9 +108,5 @@ function UnauthenticatedApp() {
   )
 }
 
-// 🐨 change this to a default export
-export {UnauthenticatedApp}
+export default UnauthenticatedApp
 
-// 🐨 Unfortunately, to make this work for our workshop,
-// you need to add this to src/unauthenticated-app.js:
-// export {default} from './unauthenticated-app.exercise'
