+++ src/unauthenticated-app.js
@@ -1,4 +1,5 @@
-export * from './unauthenticated-app.final'
-export {default} from './unauthenticated-app.final'
+// export * from './unauthenticated-app.final'
+// export {default} from './unauthenticated-app.final'
 
-// export * from './unauthenticated-app.exercise'
+export * from './unauthenticated-app.exercise'
+export {default} from './unauthenticated-app.exercise'
